requirement.txt
Pillow
numpy
python-dotenv
msedge-selenium-tools==3.141.2
python-dateutil==2.8.0
selenium==3.141.0